fix: migrate to design
