+++ src/requireAuthentication.ts
@@ -1,31 +1,31 @@
 import { GetServerSidePropsContext } from "next";
-import { unstable_getServerSession } from "next-auth/next";
-import { authOptions } from "../pages/api/auth/[...nextauth]";
-import { Profile } from "./services/AuthenticationService";
+// import { unstable_getServerSession } from "next-auth/next";
+// import { authOptions } from "../pages/api/auth/[...nextauth]";
+// import { Profile } from "./services/AuthenticationService";
 
-type Callback = (accessToken: string, user: Profile) => Promise<any>;
+// type Callback = (accessToken: string, user: Profile) => Promise<any>;
 
-const secret = process.env.NEXTAUTH_SECRET;
+// const secret = process.env.NEXTAUTH_SECRET;
 
-export const requireAuthentication = async (
-  ctx: GetServerSidePropsContext<any, any>,
-  cb: Callback
-) => {
-  const session = await unstable_getServerSession(
-    ctx.req,
-    ctx.res,
-    authOptions
-  );
+// export const requireAuthentication = async (
+//   ctx: GetServerSidePropsContext<any, any>,
+//   cb: Callback
+// ) => {
+//   const session = await unstable_getServerSession(
+//     ctx.req,
+//     ctx.res,
+//     authOptions
+//   );
 
-  if (!session) {
-    return {
-      redirect: {
-        destination: "/user/signin",
-        permanent: false,
-      },
-    };
-  }
+//   if (!session) {
+//     return {
+//       redirect: {
+//         destination: "/user/signin",
+//         permanent: false,
+//       },
+//     };
+//   }
 
-  let userSession = session as any;
-  return cb(userSession.accessToken, userSession.user);
-};
+//   let userSession = session as any;
+//   return cb(userSession.accessToken, userSession.user);
+// };
